First iteration of semantic mapping is done
In every frame, if there's an object in the rgb frame I grab all the points in the
point cloud associated with the corresponding depth image. Then I can
make smaller, sub point clouds of the points in that image.

Right now, the objects in the environment are duplicated. I'm going to
have to try and find the l2 distances between the centroids of some of
these clouds, and if it's close enough I can try and combind them.