this time, the fix is real.
The issue with objects being placed in the wrong locations in the map,
and having duplicate objects, is finally fixed. I was using a point
cloud to make the point to pixel maps that was before filtering and some
normalization, so the indexes it got were crap. on top of that, i was
accessing elements in my map with [] instead of .at(), which meant that
when trying to access an element that was out of bounds, no error would
be throw even though it was an invalid index.